contributors/rodneyrehm/submitted/css3-transitions/changing-while-transition.html

<!DOCTYPE html>
<html>
    <head>
        <meta charset="utf-8">
    	<meta http-equiv="content-type" content="text/html; charset=utf-8">
        <title>CSS Test: behavior when transition paramters change while transitioning</title>
        <meta name="assert" content="Test checks that all transitions run properly when property or duration is changed mid-run">
        <link rel="help" href="http://www.w3.org/TR/css3-transitions/#starting">
        <link rel="author" title="Rodney Rehm" href="http://rodneyrehm.de/en/">
        <meta name="flags" content="dom">
        
        <script src="/resources/testharness.js" type="text/javascript"></script>
        <script src="/resources/testharnessreport.js" type="text/javascript"></script>
        
        <script src="./support/helper.js" type="text/javascript"></script>
        <script src="./support/runAsyncHarness.js" type="text/javascript"></script>
        <script src="./support/generalTest.js" type="text/javascript"></script>
        <script src="./support/properties.js" type="text/javascript"></script>
    </head>
    <body>
        <!-- required by testharnessreport.js -->
        <div id="log"></div>
        <!-- elements used for testing -->
        <div id="fixture">
            <div id="container">
                <div id="transition">Hello World</div>
            </div>
        </div>

        <script>
            // Once the transition of a property has started, it must continue running based on the original timing function, 
            // duration, and delay, even if the ‘transition-timing-function’, ‘transition-duration’, or ‘transition-delay’ 
            // property changes before the transition is complete. However, if the ‘transition-property’ property changes 
            // such that the transition would not have started, the transition must stop (and the property must immediately 
            // change to its final value).
        
            // this test takes its time, give it 5 minutes to run
            setup({timeout: 300000});
            // init dom
            domFixture('#fixture');
            
            var tests = [
                {
                    name: "changing transition-duration",
                    transitions: 'padding-left .5s linear 0s',
                    from: {'padding-left': '1px'},
                    to: {'padding-left': '10px'},
                    then: {
                        'transition-duration': '0.1s'
                    },
                    expect: [
                        'padding-left:0.5s' 
                    ]
                },
                {
                    name: "changing transition-property",
                    transitions: 'padding-left .5s linear 0s',
                    from: {'padding-left': '1px'},
                    to: {'padding-left': '10px'},
                    then: {
                        'transition-property': 'margin-left'
                    },
                    expect: [
                        '' 
                    ]
                }
            ];
            
            // missing support for requestAnimationFrame forces us to
            // run animations at a slower pace, so setTimeout has a
            // chance of being invoked before TransitionEnd
            //var duration = getRequestAnimationFrame() ? '0.1s' : '0.5s';
            var duration = '0.5s';

            runAsyncHarness({
                // array of test data
                tests: tests,
                // milliseconds to wait before calling teardown and ending test
                duration: parseFloat(duration) * 1000,
                // prepare individual test
                setup: generalTest.setup,
                // cleanup after individual test
                teardown: generalTest.teardown,
                // run actual test, assertions can be used here!
                test: function(data, options) {
                    var styles = {
                        '#fixture': {},

                        '#container': {},
                        '#container.to': {},
                        '#container.transition': {},
                        
                        '#transition': data.from,
                        '#transition.to' : data.to,
                        '#transition.transition' : {transition: data.transitions},
                        '#transition.then' : data.then
                    };
                    
                    // prepare styles
                    setStyle(styles);
                    // kick off the transition
                    generalTest.startTransition(data);
                    setTimeout(function() {
                        data.transition.elem.classList.add('then');

                        for (var property in data.then) {
                            var current = data.transition.computedStyle(property);
                            assert_equals(current, data.then[property], 'value of ' + property + ' changed');
                        }
                        
                    }, 50);
                },
                // run tests immediately before teardown, assertions can be used here!
                concluding: function(data) {
                    // make sure we got the event for the tested property only
                    data.test.step(generalTest.assertExpectedEventsFunc(data, 'transition', data.expect));
                },
                
                // called once all tests are done
                done: generalTest.done
            });
        </script>
    </body>
</html>
